Further split source into logical units
