+++ src/piece.py
@@ -9,13 +9,11 @@ Chess has 6 types of pieces with predifined rules of movement and worth in-game:
     Rook: Moves along ranks and files. Worth 5.
     Queen: Moves both like a rook and bishop. Worth 9.
     King: Moves one square in any direction. Worth N/A (you loose the king it's game over).
-
-Here you will find the implementation of pawns as they are a special type of pieces.
 """
 
 from dataclasses import dataclass
 from enum import IntEnum
-from itertools import combinations, product
+from itertools import combinations
 from typing import ClassVar, Optional
 
 from .square import Square, Vector
@@ -189,287 +187,3 @@ class Piece:
         """
         squares, targets = set(), set()
         return squares, targets  # A ghost piece cannot move or capture.
-
-
-@dataclass(init=False, repr=False)
-class Pawn(Piece):
-    """A Pawn.
-
-    Moves forward a square, unless its first move where it can move two and unless blocked by any piece.
-    It captured forward-diagonally. I can capture en-passant:
-        If an enemy pawn passes yours trying to escape capture, well... it cannot, not for this round.
-    NOTE: That probably means that board will invoke this method multiple times (unti a better way is thought).
-    Upon reaching the end of the board, it is promoted to an officer:
-        - Rook
-        - Bishop
-        - Knight
-        - Queen
-
-    Pawn steps:
-        black:
-            ↓ one step
-            ↓ + ↓ two step
-            ↓ + ← = ↙ capture west
-            ↓ + → = ↘ capture east
-        white:
-            ↑ one step
-            ↑ + ↑ two step
-            ↑ + ← = ↖ capture west
-            ↑ + → = ↗ capture east
-    """
-
-#   Pawn value:
-    value: ClassVar[int] = 1
-
-#   Pawn moves:
-    step: ClassVar[Vector] = Vector(+1, 0)  # One step.
-    captures: ClassVar[set[Vector]] = {
-        Vector(+1, -1),  # Capturing to the west.
-        Vector(+1, +1),  # Capturing to the east.
-    }
-
-#   Pawn flags:
-    has_moved: bool = False
-    promoted: bool = False
-
-    def __repr__(self) -> str:
-        self.__repr__.__doc__
-        return {
-            "white": "♙",
-            "black": "♟",
-        }[self.orientation.name]
-
-    def moves(self) -> tuple[set[Square], set[Square]]:
-        super().moves.__doc__
-        squares, targets = super().moves()
-
-        if self.square is not None:  # If pawn is on a board,
-            for step in self.captures:  # For all target squares (diagonally with respect to pawn),
-                target = self.square + step * self.orientation  # Get target,
-
-                if target is not None:  # If said target is inside board limits,
-                    targets.add(target)  # Add said target to pawn.
-
-            square = self.square + self.step * self.orientation  # Get forward square,
-
-            if square is None:  # If said square is outside board limits,
-                self.promoted = True  # The pawn is promoted, assuming it falls outside because it can only move forward.
-
-            else:  # Otherwise,
-                squares.add(square)  # Add said square to possible moves,
-
-                if not self.has_moved:  # If the pawn is in its starting position,
-                    squares.add(square + self.step * self.orientation)  # Add the next forward square to possible moves too.
-
-        return squares, targets
-
-
-@dataclass(init=False, repr=False)
-class Meleed(Piece):
-    """A close ranged piece:
-        - King
-        - Knight (has reach)
-
-    For a meeled piece (pawns are special) all squares are target squares.
-    """
-
-    def moves(self) -> tuple[set[Square], set[Square]]:
-        super().moves.__doc__
-        squares, targets = super().moves()
-
-        if self.square is not None:  # If meleed piece is on a board,
-            for step in self.steps:  # For all target squares,
-                target = self.square + step  # Get target,
-
-                if target is not None and self.condition(target):  # If said target is inside board limits,
-                    targets.add(target)  # Add said target to meleed piece.
-
-        return squares, targets
-
-
-@dataclass(init=False, repr=False)
-class King(Meleed):
-    """A King.
-
-    Moves one square in any direction that is not blocked by same side piece or targeted by enemy piece.
-
-    King steps:
-        ↑ north
-        ← west
-        ↓ south
-        → east
-        ↗ north-east
-        ↘ south-east
-        ↖ north-west
-        ↙ south-west
-    """
-
-#   King moves:
-    steps: ClassVar[set[Vector]] = Piece.straights | Piece.diagonals
-
-    def __repr__(self) -> str:
-        super().__repr__.__doc__
-        return {
-            "white": "♔",
-            "black": "♚",
-        }[self.orientation.name]
-
-
-@dataclass(init=False, repr=False)
-class Knight(Meleed):
-    """A knight.
-
-    Moves two squares in any direction and one square in a vertical to that direction as one move.
-    It therefore can jump over other pieces. Target square must not be blocked by a same side piece.
-
-    Knight steps (via product of diagonal steps with straigh steps:
-        ↗ + ↑ north-north-east
-        ↗ + ← = ↑ north (removed)
-        ↗ + ↓ = → east (removed)
-        ↗ + → north-east-east
-        ↘ + ↑ = → east (removed)
-        ↘ + ← = ↓ south (removed)
-        ↘ + ↓ south-south-east
-        ↘ + → south-east-east
-        ↖ + ↑ north-north-west
-        ↖ + ← north-west-west
-        ↖ + ↓ = ← west (removed)
-        ↖ + → = ↑ north (removed)
-        ↙ + ↑ = ← west (removed)
-        ↙ + ← south-west-west
-        ↙ + ↓ south-south-west
-        ↙ + → = ↓ south (removed)
-    """
-
-#   Knight value:
-    value: ClassVar[int] = 3
-
-#   Knight moves:
-    steps: ClassVar[set[Vector]] = \
-        {diagonal + straight for diagonal, straight in product(Piece.diagonals, Piece.straights)} - Piece.straights
-
-    def __repr__(self) -> str:
-        super().__repr__.__doc__
-        return {
-            "white": "♘",
-            "black": "♞",
-        }[self.orientation.name]
-
-
-@dataclass(init=False, repr=False)
-class Ranged(Piece):
-    """A long range piece:
-        - Rook
-        - Bishop
-        - Queen
-    """
-
-    def moves(self) -> tuple[set[Square], set[Square]]:
-        super().moves.__doc__
-        squares, targets = super().moves()
-
-        if self.square is not None:  # If ranged piece is on a board,
-            for step in self.steps:  # For all legal directions,
-                square = self.square + step  # Get next square in direction,
-
-                while square is not None and self.condition(square):  # If said square is inside board limits,
-                    squares.add(square)  # Add said square to ranged piece.
-                    square += step  # Advance to the next square in said direction.
-
-                if square is not None:  # NOTE: This check will become relevant as soon as condition is defined too.
-                    targets.add(square)  # Do not forget to add trailing square to targets.
-
-        return squares, targets
-
-
-@dataclass(init=False, repr=False)
-class Rook(Ranged):
-    """A rook.
-
-    Moves anywhere along its rank xor its file, unless blocked by enemy (which captures) or same piece.
-
-    Rook steps:
-        ↑ north
-        ← west
-        ↓ south
-        → east
-    """
-
-#   Knight value:
-    value: ClassVar[int] = 5
-
-#   Straight lines:
-    steps: ClassVar[set[Vector]] = Piece.straights
-
-    def __repr__(self) -> str:
-        super().__repr__.__doc__
-        return {
-            "white": "♖",
-            "black": "♜",
-        }[self.orientation.name]
-
-
-@dataclass(init=False, repr=False)
-class Bishop(Ranged):
-    """A bishop.
-
-    Moves anywhere along a diagonal, unless blocked by enemy (which captures) or same piece.
-    This means that a bishop never leaves the color of the square it is on.
-    A player starts with two bishops:
-        A black-square bishop.
-        A white-square bishop.
-    So both cover the squares of the whole board, and it is the reason the worth more when both on the board.
-
-    Bishop steps:
-        ↗ north-east
-        ↘ south-east
-        ↖ north-west
-        ↙ south-west
-    """
-
-#   Knight value:
-    value: ClassVar[int] = 3
-
-#   Diagonal lines:
-    steps: ClassVar[set[Vector]] = Piece.diagonals
-
-    def __repr__(self) -> str:
-        super().__repr__.__doc__
-        return {
-            "white": "♗",
-            "black": "♝",
-        }[self.orientation.name]
-
-
-@dataclass(init=False, repr=False)
-class Queen(Ranged):
-    """A queen.
-
-    Moves both like a rook and bishop, unless blocked by enemy (which captures) or same piece.
-    Can access all squares.
-
-    Inherits the steps of a King but the movement of Range.
-
-    Queen steps:
-        ↑ north
-        ← west
-        ↓ south
-        → east
-        ↗ north-east
-        ↘ south-east
-        ↖ north-west
-        ↙ south-west
-    """
-
-#   Knight value:
-    value: ClassVar[int] = 9
-
-#   Queen lines:
-    steps: ClassVar[set[Vector]] = Piece.straights | Piece.diagonals
-
-    def __repr__(self) -> str:
-        super().__repr__.__doc__
-        return {
-            "white": "♕",
-            "black": "♛",
-        }[self.orientation.name]
