+++ test_chess.py
@@ -64,7 +64,9 @@ class TestPiece:
 
     def test_legal_moves(self):
         """Test whether the pieces generate proper legal moves. One example for each piece."""
-        from src.piece import Bishop, King, Knight, Pawn, Queen, Rook
+        from src.pieces.meleed import King, Knight
+        from src.pieces.ranged import Bishop, Queen, Rook
+        from src.pieces.special import Pawn
         from src.square import Square
 
         assert Pawn("white", Square("h2")).moves() == (
